Clean up webpack code
