+++ package.json
@@ -4,20 +4,13 @@
   "description": "A viewer to display codeclub lessons",
   "main": "index.js",
   "scripts": {
-    "start": "webpack-dev-server --config=webpack.new.config.babel.js -d --progress --inline --hot --history-api-fallback --host 0.0.0.0",
-    "build:debug": "webpack --config=webpack.new.config.babel.js -d --progress --env.verbose",
-    "build": "webpack --config=webpack.new.config.babel.js --progress",
-    "build:prod": "webpack --config=webpack.new.config.babel.js --progress",
-    "build:analyse:": "webpack --config=webpack.new.config.babel.js --progress --profile --json > stats.json",
-    "buildjs_old": "webpack -d --progress",
-    "buildjs_old:prod": "cross-env NODE_ENV=production BUILD_PDF=true webpack -p --progress",
-    "buildstatic_old": "webpack --config=webpack.static.config.babel.js --progress",
-    "buildstatic_old:prod": "cross-env BUILD_PDF=true webpack --config=webpack.static.config.babel.js --progress",
-    "build_old": "yarn buildjs_old && yarn buildstatic_old",
-    "build_old:prod": "yarn buildjs_old:prod && yarn buildstatic_old:prod",
-    "build_old:travis": "cross-env NODE_ENV=production webpack -p && webpack --config=webpack.static.config.babel.js",
+    "start": "webpack-dev-server -d --progress --inline --hot --history-api-fallback --host 0.0.0.0",
+    "build:debug": "webpack -d --progress --env.verbose",
+    "build": "webpack --progress",
+    "build:prod": "webpack -p --progress --env.NODE_ENV=production --env.BUILD_PDF",
+    "build:analyse:": "webpack --progress --profile --json > stats.json",
+    "build:travis": "webpack -p --env.NODE_ENV=production",
     "serve": "ws",
-    "servejs": "node ./server.js",
     "test": "mocha --compilers js:babel-core/register --require mock-local-storage --require ignore-styles --require ./test/test_helper.js \"test/**/*@(.js|.jsx)\"",
     "test:links": "node test-links.js",
     "test:watch": "yarn test -- --watch",
